Move "Dumb" test objects to base file
Tests should not directly use abstract classes, but now use basic implementation instead
